% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictionErrorGMM.R
\name{predictionErrorGMM}
\alias{predictionErrorGMM}
\title{Generalized method of moments (GMM) estimator for overcoming
prediction error}
\usage{
predictionErrorGMM(Yl, Xl, Zl, Vl = NULL, Yu, Zu, Vu = NULL,
  Yt = NULL, Xt = NULL, Vt = NULL, num_boot = 1000)
}
\arguments{
\item{Yl}{vector of outcome values for the mapping sample}

\item{Xl}{vector of X covariate values for the mapping sample}

\item{Zl}{matrix of Z predicted covariates values for the mapping sample}

\item{Vl}{matrix of V other covariate values for the mapping sample}

\item{Yu}{vector of outcome values for the unlabeled sample}

\item{Zu}{matrix of Z predicted covariates values for the unlabeled sample}

\item{Vu}{matrix of V other covariate values for the unlabeled sample}

\item{Yt}{vector of outcome values for the training sample (optional)}

\item{Xt}{matrix of X covariates values for the training sample (optional)}

\item{Vt}{matrix of V other covariate values for the training sample (optional)}

\item{num_boot}{number of bootstrap iterations to use in bootstrap asymptotic refinement; set to zero if you do not want to bootstrap the exclusion restriciton test p-value}
}
\value{
A list of
\itemize{
\item{estimated coefficients \code{coefficients}}
\item{estimated standard errors \code{stderr}}
\item{covariance matrix estimate \code{vcov}}
\item{estimated coefficient of the residual on Z \code{rho_est} to evalute direction of exclusion restriction violation}
\item{exclusion restriction (overidentification) test statistic \code{OI_test_stat}}
\item{conventional chi-squared overidentificaiton test p-value \code{OI_test_pval_chisq}}
\item{bootstrap with asymptotic refinement overidentificaiton test p-value \code{OI_test_pval_boot}}
}
}
\description{
Combines ordinary least squares (OLS) on the hand-labeled set and
two-sample two-stage least squares (TS2SLS) on the unlabeled
set. See associated paper for details.
}
